pyproject.toml
[project]
name = "digit-classification"
version = "0.1.0"
description = "Solution to the digit classification problem."
readme = "README.md"
authors = [
    { name = "Shudipto Sekhar Roy", email = "shudipto67@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "lightning>=2.5.0.post0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "typer>=0.15.0",
    "torchmetrics>=1.4.0",
    "scikit-learn>=1.3.0",
    "numpy==2.3.4",
    "pillow>=10.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0"
]
dev = [
    "black>=24.0.0",
    "flake8>=7.0.0"
]

[project.scripts]
digit-classification = "digit_classification.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/digit_classification"]